requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
google-generativeai==0.3.2
yt-dlp==2023.11.16
ffmpeg-python==0.2.0
jinja2==3.1.3
python-dotenv==1.0.1
requests==2.31.0
